requirements.txt
flask==2.2.5
gunicorn==20.1.0
flask-socketio==5.3.6
eventlet==0.36.1
python-dotenv==1.0.1

